Refactor: Replace `Emitter` class with `AbstractEmitter`, simplifying and centralizing emitter logic
