+++ src/compiler/emitters/kotlin/src/commonMain/kotlin/community/flock/wirespec/emitters/kotlin/KotlinEndpointDefinitionEmitter.kt
@@ -1,8 +1,8 @@
 package community.flock.wirespec.emitters.kotlin
 
 import community.flock.wirespec.compiler.core.concatGenerics
-import community.flock.wirespec.compiler.core.emit.Emitter.Companion.firstToLower
-import community.flock.wirespec.compiler.core.emit.Emitter.Companion.isStatusCode
+import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToLower
+import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.isStatusCode
 import community.flock.wirespec.compiler.core.emit.EndpointDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.EndpointEmitter
 import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
