+++ src/compiler/emitters/typescript/src/commonMain/kotlin/community/flock/wirespec/emitters/typescript/TypeScriptEmitter.kt
@@ -2,7 +2,7 @@ package community.flock.wirespec.emitters.typescript
 
 import arrow.core.NonEmptyList
 import community.flock.wirespec.compiler.core.emit.Emitted
-import community.flock.wirespec.compiler.core.emit.Emitter
+import community.flock.wirespec.compiler.core.emit.AbstractEmitter
 import community.flock.wirespec.compiler.core.emit.FileExtension
 import community.flock.wirespec.compiler.core.emit.PackageName
 import community.flock.wirespec.compiler.core.emit.namespace
@@ -14,7 +14,7 @@ import community.flock.wirespec.compiler.core.parse.Module
 import community.flock.wirespec.compiler.core.parse.Union
 import community.flock.wirespec.compiler.utils.Logger
 
-open class TypeScriptEmitter() : Emitter(), TypeScriptEnumDefinitionEmitter, TypeScriptIdentifierEmitter, TypeScriptTypeDefinitionEmitter, TypeScriptClientEmitter, TypeScriptEndpointDefinitionEmitter, TypeScriptRefinedTypeDefinitionEmitter {
+open class TypeScriptEmitter() : AbstractEmitter(), TypeScriptEnumDefinitionEmitter, TypeScriptIdentifierEmitter, TypeScriptTypeDefinitionEmitter, TypeScriptClientEmitter, TypeScriptEndpointDefinitionEmitter, TypeScriptRefinedTypeDefinitionEmitter {
 
     override val extension = FileExtension.TypeScript
 
@@ -22,7 +22,7 @@ open class TypeScriptEmitter() : Emitter(), TypeScriptEnumDefinitionEmitter, Typ
 
     override val singleLineComment = "//"
 
-    override fun emit(ast: AST, logger: Logger): NonEmptyList<Emitted> = super<Emitter>.emit(ast, logger)
+    override fun emit(ast: AST, logger: Logger): NonEmptyList<Emitted> = super<AbstractEmitter>.emit(ast, logger)
         .run { if(ast.hasEndpoints()){ plus(emitClient(ast) ) } else this }
         .plus(
             ast.modules
@@ -36,12 +36,12 @@ open class TypeScriptEmitter() : Emitter(), TypeScriptEnumDefinitionEmitter, Typ
                 }
         )
 
-    override fun emit(module: Module, logger: Logger): NonEmptyList<Emitted> = super<Emitter>.emit(module, logger).let {
+    override fun emit(module: Module, logger: Logger): NonEmptyList<Emitted> = super<AbstractEmitter>.emit(module, logger).let {
         it + Emitted("Wirespec", shared.source)
     }
 
     override fun emit(definition: Definition, module: Module, logger: Logger): Emitted =
-        super<Emitter>.emit(definition, module, logger).let {
+        super<AbstractEmitter>.emit(definition, module, logger).let {
             val subPackageName = PackageName("") + definition
             Emitted(
                 file = subPackageName.toDir() + it.file.sanitizeSymbol(),
