+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecEmitter.kt
@@ -1,22 +1,9 @@
 package community.flock.wirespec.emitters.wirespec
 
-import community.flock.wirespec.compiler.core.addBackticks
-import community.flock.wirespec.compiler.core.emit.Emitter
+import community.flock.wirespec.compiler.core.emit.AbstractEmitter
 import community.flock.wirespec.compiler.core.emit.FileExtension
 import community.flock.wirespec.compiler.core.emit.Keywords
-import community.flock.wirespec.compiler.core.emit.Spacer
-import community.flock.wirespec.compiler.core.parse.Channel
-import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
-import community.flock.wirespec.compiler.core.parse.Endpoint
-import community.flock.wirespec.compiler.core.parse.Enum
-import community.flock.wirespec.compiler.core.parse.Field
-import community.flock.wirespec.compiler.core.parse.FieldIdentifier
-import community.flock.wirespec.compiler.core.parse.Identifier
-import community.flock.wirespec.compiler.core.parse.Module
 import community.flock.wirespec.compiler.core.parse.Reference
-import community.flock.wirespec.compiler.core.parse.Refined
-import community.flock.wirespec.compiler.core.parse.Type
-import community.flock.wirespec.compiler.core.parse.Union
 
 interface E:
     WirespecIdentifierEmitter,
@@ -27,7 +14,7 @@ interface E:
     WirespecUnionDefinitionEmitter,
     WirespecRefinedTypeDefinitionEmitter
 
-open class WirespecEmitter : Emitter(), E {
+open class WirespecEmitter : AbstractEmitter(), E {
 
     override val extension = FileExtension.Wirespec
