+++ src/compiler/test/src/commonMain/kotlin/community/flock/wirespec/compiler/test/TestUtils.kt
@@ -6,10 +6,10 @@ import community.flock.wirespec.compiler.core.CompilationContext
 import community.flock.wirespec.compiler.core.ModuleContent
 import community.flock.wirespec.compiler.core.WirespecSpec
 import community.flock.wirespec.compiler.core.compile
-import community.flock.wirespec.compiler.core.emit.Emitter
+import community.flock.wirespec.compiler.core.emit.AbstractEmitter
 import community.flock.wirespec.compiler.utils.NoLogger
 
-fun compile(source: String) = { emitter: () -> Emitter ->
+fun compile(source: String) = { emitter: () -> AbstractEmitter ->
     object : CompilationContext, NoLogger {
         override val spec = WirespecSpec
         override val emitters = nonEmptySetOf(emitter())
